// src/types/fhir/Bundle.ts
// WARNING: This file is autogenerated by FHIR Schema Codegen.
// https://github.com/fhir-schema/fhir-schema-codegen
// Any manual changes made to this file may be overwritten.

import type { BackboneElement } from './BackboneElement';
import type { Element } from './Element';
import type { Identifier } from './Identifier';
import type { Resource } from './Resource';
import type { Signature } from './Signature';


export interface BundleEntry<T> extends BackboneElement {
    fullUrl?: string;
    link?: BundleLink[];
    request?: BundleEntryRequest;
    resource?: T;
    response?: BundleEntryResponse;
    search?: BundleEntrySearch;
}

export interface BundleEntryRequest extends BackboneElement {
    ifMatch?: string;
    ifModifiedSince?: string;
    ifNoneExist?: string;
    ifNoneMatch?: string;
    method: 'GET' | 'HEAD' | 'POST' | 'PUT' | 'DELETE' | 'PATCH';
    url: string;
}

export interface BundleEntryResponse extends BackboneElement {
    etag?: string;
    lastModified?: string;
    location?: string;
    outcome?: Resource;
    status: string;
}

export interface BundleEntrySearch extends BackboneElement {
    mode?: 'match' | 'include' | 'outcome';
    score?: number;
}

export interface BundleLink extends BackboneElement {
    relation: string;
    url: string;
}

export interface Bundle<T> extends Resource {
    entry?: BundleEntry<T>[];
    identifier?: Identifier;
    link?: BundleLink[];
    signature?: Signature;
    timestamp?: string;
    _timestamp?: Element;
    total?: number;
    _total?: Element;
    type: 'document' | 'message' | 'transaction' | 'transaction-response' | 'batch' | 'batch-response' | 'history' | 'searchset' | 'collection';
    _type?: Element;
}